requirements.txt
streamlit>=1.22.0
requests>=2.28.0
google-generativeai>=0.3.0
streamlit-agraph>=0.0.7
python-dotenv>=0.19.0
